practice_004a_grpc_python/app/pyproject.toml
[project]
name = "task-manager-grpc"
version = "0.1.0"
description = "Practice 004a: gRPC & Protobuf Python Service (TaskManager)"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
]
